player
moving player class into `luaplayer.cpp/h`
